grpc-stt-server/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
faster-whisper
torch>=2.0.0
numpy
protobuf>=4.21.0